refactors victory check and folder structure
