+++ src/components/BottomContent/BetOptions.tsx
