+++ src/components/BottomContent/index.tsx
